Rename 'Value' to 'Success'
Rename the value type to success to more accurately reflect the
structures that are listed in the specification. It's a bit gnarly but
this is more faithful to the spec and should result in easier
translation and adaptation.

Updated the tests.

Tightened up some of the other types, things are starting to look pretty
good.