Added Cookie type, moved WDUri to own module
The new Cookie type is an approximation of the cookie structure
defined in the Webdriver spec, with the value of the cookie left as a
JSON structure that should provide enough flexibility for most users.

Moved the WDUri to its own module to avoid a module cycle.

Updated session types to use new Cookie type.

Deleted AddCookie type as it was better served by the boring Cookie
type. Might need a better type actually to account for required fields.