Add HollowBody combinator.
The Webdriver spec does not indicate that some POST commands require any
content, and indeed they do not "require" any content in the body of the
request. But they will fail if there is not at least an empty JSON
object.

To try to make this easier for the user because this is a silly
requirement, the `HollowBody` combinator was created. It's task is to
simply populate the body of the request with an encoded empty JSON
object to satisfy the endpoints.

Updated some routes to use the new combinator.

Rewrote the README.

Moved the `WDCore` type to the library proper as it is a nicer way of
handling the creation of the related clients for a given webdriver
execution.

Added some tests for alerts using a `window.prompt` script.

As a by product, the `executeScript` command appears to work.

Shenanigans.