Fleshed out Action type.
The `performActions` route has a picky structure that was better served
by a bunch of newtypes and related things. There are no decoders written
for this type and I'll defer that until they're really needed.

Added a unit test for pressing tab and checking active elements etc.

Added beginnings of page for an accessibility test for focus traps if I
ever get around to it.